Ajout de la fonction de lecture des données
Modification d'un bug dans la classe DonneesArgos pour l'ajout des
latitude et longitudes
Ajout d'un fichier de position
Ajout de la fonction permettant la lecture des données à partir des
fichiers de position

+++ src/argoview/DonneeArgos.java
@@ -37,8 +37,8 @@ public class DonneeArgos {
         
         this.date = new GregorianCalendar(annee, mois, jour, heures, minutes);
         
-        this.latitude  = Integer.parseInt(latitude);
-        this.longitude = Integer.parseInt(longitude);
+        this.latitude  = Double.parseDouble(latitude);
+        this.longitude = Double.parseDouble(longitude);
     }
     
     /*
